Refactor CLI Structure (#224)
* Flatten CLI

* Flatten CLI

* Refactor

* Use datastore module

* Extract main functions from password_delete
