+++ packages/ciphernode/enclave/src/helpers/compile_id.rs
