+++ packages/ciphernode/enclave_node/src/helpers/shutdown.rs
