+++ packages/ciphernode/enclave_node/src/password_delete.rs
@@ -1,0 +1,21 @@
+use anyhow::*;
+use cipher::{FilePasswordManager, PasswordManager};
+use config::AppConfig;
+use zeroize::Zeroizing;
+
+pub async fn get_current_password(config: &AppConfig) -> Result<Zeroizing<String>> {
+    let key_file = config.key_file();
+    let pm = FilePasswordManager::new(key_file);
+    if !pm.is_set() {
+        bail!("Password is not set. Nothing to do.")
+    }
+    let pw = pm.get_key().await?;
+    Ok(Zeroizing::new(String::from_utf8_lossy(&pw).to_string()))
+}
+
+pub async fn execute(config: &AppConfig) -> Result<()> {
+    let key_file = config.key_file();
+    let mut pm = FilePasswordManager::new(key_file);
+    pm.delete_key().await?;
+    Ok(())
+}
