+++ packages/ciphernode/evm/tests/evm_reader.rs
@@ -9,7 +9,7 @@ use alloy::{
 use anyhow::Result;
 use data::Repository;
 use enclave_core::{EnclaveEvent, EventBus, GetHistory, Shutdown, TestEvent};
-use enclave_node::get_in_mem_store;
+use enclave_node::datastore::get_in_mem_store;
 use evm::{helpers::WithChainId, EvmEventReader};
 use std::time::Duration;
 use tokio::time::sleep;
